Created arguments for authorizer and description
Went through the commands and made passing description as well
as the 'authorizer' available via the command line.  Hopefully
this means that in the future, passwords can be dumped into
pkpass via stdin without exposing them via command line argu-
ments.  Everything required to create a password except for
the password itself is now avaialble to input on the command
line.